+++ History.md
@@ -7,6 +7,8 @@
   * Cleanup daemonization in rc.d script (#2409)
 
 * Refactor
+  * Consolidate option handling in Server, Server small refactors, doc chang (#2389)
+  * Refactor Reactor and Client request buffering (#2279)
   * client.rb - remove JRuby specific 'finish' code (#2412)
   * Consolidate fast_write calls in Server, extract early_hints assembly (#2405)
   * Remove upstart from docs (#2408)